Reduce AI slop
